move store_statistics into statistics package
