+++ client/client_test.go
@@ -25,10 +25,10 @@ import (
 	. "github.com/pingcap/check"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/pd/pkg/mock/mockid"
 	"github.com/pingcap/pd/pkg/testutil"
 	"github.com/pingcap/pd/server"
 	"github.com/pingcap/pd/server/core"
-	"github.com/pingcap/pd/server/mock"
 	"google.golang.org/grpc"
 )
 
@@ -40,7 +40,7 @@ func TestClient(t *testing.T) {
 var _ = Suite(&testClientSuite{})
 
 type idAllocator struct {
-	allocator *mock.IDAllocator
+	allocator *mockid.IDAllocator
 }
 
 func (i *idAllocator) alloc() uint64 {
@@ -49,7 +49,7 @@ func (i *idAllocator) alloc() uint64 {
 }
 
 var (
-	regionIDAllocator = &idAllocator{allocator: &mock.IDAllocator{}}
+	regionIDAllocator = &idAllocator{allocator: &mockid.IDAllocator{}}
 	// Note: IDs below are entirely arbitrary. They are only for checking
 	// whether GetRegion/GetStore works.
 	// If we alloc ID in client in the future, these IDs must be updated.
