+++ pkg/mock/mockoption/mockoption.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mock
+package mockoption
 
 import (
 	"time"
