+++ server/coordinator_test.go
@@ -23,9 +23,10 @@ import (
 	"github.com/pingcap/kvproto/pkg/eraftpb"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/pd/pkg/mock/mockhbstream"
+	"github.com/pingcap/pd/pkg/mock/mockid"
 	"github.com/pingcap/pd/pkg/testutil"
 	"github.com/pingcap/pd/server/core"
-	"github.com/pingcap/pd/server/mock"
 	"github.com/pingcap/pd/server/namespace"
 	"github.com/pingcap/pd/server/schedule"
 	"github.com/pingcap/pd/server/schedulers"
@@ -52,7 +53,7 @@ type testClusterInfo struct {
 
 func newTestClusterInfo(opt *scheduleOption) *testClusterInfo {
 	return &testClusterInfo{clusterInfo: newClusterInfo(
-		mock.NewIDAllocator(),
+		mockid.NewIDAllocator(),
 		opt,
 		core.NewKV(core.NewMemoryKV()),
 	)}
@@ -205,7 +206,7 @@ func (s *testCoordinatorSuite) TestDispatch(c *C) {
 	testutil.CheckTransferLeader(c, co.opController.GetOperator(2), schedule.OpBalance, 4, 2)
 	c.Assert(co.removeScheduler("balance-leader-scheduler"), IsNil)
 
-	stream := mock.NewHeartbeatStream()
+	stream := mockhbstream.NewHeartbeatStream()
 
 	// Transfer peer.
 	region := tc.GetRegion(1).Clone()
@@ -227,7 +228,7 @@ func (s *testCoordinatorSuite) TestDispatch(c *C) {
 	waitNoResponse(c, stream)
 }
 
-func dispatchHeartbeat(c *C, co *coordinator, region *core.RegionInfo, stream mock.HeartbeatStream) error {
+func dispatchHeartbeat(c *C, co *coordinator, region *core.RegionInfo, stream mockhbstream.HeartbeatStream) error {
 	co.hbStreams.bindStream(region.GetLeader().GetStoreId(), stream)
 	if err := co.cluster.putRegion(region.Clone()); err != nil {
 		return err
@@ -337,7 +338,7 @@ func (s *testCoordinatorSuite) TestReplica(c *C) {
 	c.Assert(tc.addRegionStore(3, 3), IsNil)
 	c.Assert(tc.addRegionStore(4, 4), IsNil)
 
-	stream := mock.NewHeartbeatStream()
+	stream := mockhbstream.NewHeartbeatStream()
 
 	// Add peer to store 1.
 	c.Assert(tc.addLeaderRegion(1, 2, 3), IsNil)
@@ -402,7 +403,7 @@ func (s *testCoordinatorSuite) TestPeerState(c *C) {
 	c.Assert(tc.addRegionStore(4, 40), IsNil)
 	c.Assert(tc.addLeaderRegion(1, 2, 3, 4), IsNil)
 
-	stream := mock.NewHeartbeatStream()
+	stream := mockhbstream.NewHeartbeatStream()
 
 	// Wait for schedule.
 	waitOperator(c, co, 1)
@@ -550,7 +551,7 @@ func (s *testCoordinatorSuite) TestAddScheduler(c *C) {
 	c.Assert(co.removeScheduler("label-scheduler"), IsNil)
 	c.Assert(co.schedulers, HasLen, 0)
 
-	stream := mock.NewHeartbeatStream()
+	stream := mockhbstream.NewHeartbeatStream()
 
 	// Add stores 1,2,3
 	c.Assert(tc.addLeaderStore(1, 1), IsNil)
@@ -703,7 +704,7 @@ func (s *testCoordinatorSuite) TestRestart(c *C) {
 	// Add 1 replica on store 2.
 	co := newCoordinator(tc.clusterInfo, hbStreams, namespace.DefaultClassifier)
 	co.run()
-	stream := mock.NewHeartbeatStream()
+	stream := mockhbstream.NewHeartbeatStream()
 	c.Assert(dispatchHeartbeat(c, co, region, stream), IsNil)
 	region = waitAddLearner(c, stream, region, 2)
 	c.Assert(dispatchHeartbeat(c, co, region, stream), IsNil)
@@ -736,7 +737,7 @@ func (s *testOperatorControllerSuite) TestOperatorCount(c *C) {
 	_, opt, err := newTestScheduleConfig()
 	c.Assert(err, IsNil)
 	tc := newTestClusterInfo(opt)
-	hbStreams := mock.NewHeartbeatStreams(tc.clusterInfo.getClusterID())
+	hbStreams := mockhbstream.NewHeartbeatStreams(tc.clusterInfo.getClusterID())
 
 	oc := schedule.NewOperatorController(tc.clusterInfo, hbStreams)
 	c.Assert(oc.OperatorCount(schedule.OpLeader), Equals, uint64(0))
@@ -881,7 +882,7 @@ func (s *testScheduleControllerSuite) TestInterval(c *C) {
 	}
 }
 
-func waitAddLearner(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+func waitAddLearner(c *C, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
 	var res *pdpb.RegionHeartbeatResponse
 	testutil.WaitUntil(c, func(c *C) bool {
 		if res = stream.Recv(); res != nil {
@@ -897,7 +898,7 @@ func waitAddLearner(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, 
 	)
 }
 
-func waitPromoteLearner(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+func waitPromoteLearner(c *C, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
 	var res *pdpb.RegionHeartbeatResponse
 	testutil.WaitUntil(c, func(c *C) bool {
 		if res = stream.Recv(); res != nil {
@@ -914,7 +915,7 @@ func waitPromoteLearner(c *C, stream mock.HeartbeatStream, region *core.RegionIn
 	)
 }
 
-func waitRemovePeer(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+func waitRemovePeer(c *C, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
 	var res *pdpb.RegionHeartbeatResponse
 	testutil.WaitUntil(c, func(c *C) bool {
 		if res = stream.Recv(); res != nil {
@@ -930,7 +931,7 @@ func waitRemovePeer(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, 
 	)
 }
 
-func waitTransferLeader(c *C, stream mock.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+func waitTransferLeader(c *C, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
 	var res *pdpb.RegionHeartbeatResponse
 	testutil.WaitUntil(c, func(c *C) bool {
 		if res = stream.Recv(); res != nil {
@@ -943,7 +944,7 @@ func waitTransferLeader(c *C, stream mock.HeartbeatStream, region *core.RegionIn
 	)
 }
 
-func waitNoResponse(c *C, stream mock.HeartbeatStream) {
+func waitNoResponse(c *C, stream mockhbstream.HeartbeatStream) {
 	testutil.WaitUntil(c, func(c *C) bool {
 		res := stream.Recv()
 		return res == nil
