+++ server/schedulers/scheduler_test.go
@@ -16,9 +16,11 @@ package schedulers
 import (
 	. "github.com/pingcap/check"
 	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/pingcap/pd/pkg/mock/mockcluster"
+	"github.com/pingcap/pd/pkg/mock/mockhbstream"
+	"github.com/pingcap/pd/pkg/mock/mockoption"
 	"github.com/pingcap/pd/pkg/testutil"
 	"github.com/pingcap/pd/server/core"
-	"github.com/pingcap/pd/server/mock"
 	"github.com/pingcap/pd/server/namespace"
 	"github.com/pingcap/pd/server/schedule"
 	"github.com/pingcap/pd/server/statistics"
@@ -29,8 +31,8 @@ var _ = Suite(&testShuffleLeaderSuite{})
 type testShuffleLeaderSuite struct{}
 
 func (s *testShuffleLeaderSuite) TestShuffle(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	sl, err := schedule.CreateScheduler("shuffle-leader", schedule.NewOperatorController(nil, nil))
 	c.Assert(err, IsNil)
@@ -59,8 +61,8 @@ var _ = Suite(&testBalanceAdjacentRegionSuite{})
 type testBalanceAdjacentRegionSuite struct{}
 
 func (s *testBalanceAdjacentRegionSuite) TestBalance(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	sc, err := schedule.CreateScheduler("adjacent-region", schedule.NewOperatorController(nil, nil), "32", "2")
 	c.Assert(err, IsNil)
@@ -127,8 +129,8 @@ func (s *testBalanceAdjacentRegionSuite) TestBalance(c *C) {
 }
 
 func (s *testBalanceAdjacentRegionSuite) TestNoNeedToBalance(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	sc, err := schedule.CreateScheduler("adjacent-region", schedule.NewOperatorController(nil, nil))
 	c.Assert(err, IsNil)
@@ -181,8 +183,8 @@ func (s *testScatterRegionSuite) checkOperator(op *schedule.Operator, c *C) {
 }
 
 func (s *testScatterRegionSuite) scatter(c *C, numStores, numRegions uint64) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	// Add stores 1~6.
 	for i := uint64(1); i <= numStores; i++ {
@@ -222,9 +224,9 @@ func (s *testScatterRegionSuite) scatter(c *C, numStores, numRegions uint64) {
 }
 
 func (s *testScatterRegionSuite) TestStorelimit(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
-	oc := schedule.NewOperatorController(tc, mock.NewHeartbeatStream())
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
+	oc := schedule.NewOperatorController(tc, mockhbstream.NewHeartbeatStream())
 
 	// Add stores 1~6.
 	for i := uint64(1); i <= 5; i++ {
@@ -254,11 +256,11 @@ var _ = Suite(&testRejectLeaderSuite{})
 type testRejectLeaderSuite struct{}
 
 func (s *testRejectLeaderSuite) TestRejectLeader(c *C) {
-	opt := mock.NewScheduleOptions()
+	opt := mockoption.NewScheduleOptions()
 	opt.LabelProperties = map[string][]*metapb.StoreLabel{
 		schedule.RejectLeader: {{Key: "noleader", Value: "true"}},
 	}
-	tc := mock.NewCluster(opt)
+	tc := mockcluster.NewCluster(opt)
 
 	// Add 3 stores 1,2,3.
 	tc.AddLabelsStore(1, 1, map[string]string{"noleader": "true"})
@@ -313,9 +315,9 @@ var _ = Suite(&testShuffleHotRegionSchedulerSuite{})
 type testShuffleHotRegionSchedulerSuite struct{}
 
 func (s *testShuffleHotRegionSchedulerSuite) TestBalance(c *C) {
-	opt := mock.NewScheduleOptions()
+	opt := mockoption.NewScheduleOptions()
 	newTestReplication(opt, 3, "zone", "host")
-	tc := mock.NewCluster(opt)
+	tc := mockcluster.NewCluster(opt)
 	hb, err := schedule.CreateScheduler("shuffle-hot-region", schedule.NewOperatorController(nil, nil))
 	c.Assert(err, IsNil)
 
@@ -364,8 +366,8 @@ var _ = Suite(&testEvictLeaderSuite{})
 type testEvictLeaderSuite struct{}
 
 func (s *testEvictLeaderSuite) TestEvictLeader(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	// Add stores 1, 2, 3
 	tc.AddLeaderStore(1, 0)
@@ -388,8 +390,8 @@ var _ = Suite(&testShuffleRegionSuite{})
 type testShuffleRegionSuite struct{}
 
 func (s *testShuffleRegionSuite) TestShuffle(c *C) {
-	opt := mock.NewScheduleOptions()
-	tc := mock.NewCluster(opt)
+	opt := mockoption.NewScheduleOptions()
+	tc := mockcluster.NewCluster(opt)
 
 	sl, err := schedule.CreateScheduler("shuffle-region", schedule.NewOperatorController(nil, nil))
 	c.Assert(err, IsNil)
